New file for SIMD/jemalloc/mimalloc settings
